insights agent/ai_insights_requirements.txt
openai>=1.65.0
google-genai>=1.10.0
requests>=2.31.0
